pyproject.toml
[tool.poetry]
name = "test-pyenv-poetry"
version = "0.1.0"
description = ""
authors = ["Marsel <lllmarselll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
flake8 = "^7.1.1"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
